+++ dandy/bot/__init__.py
@@ -1,2 +1,1 @@
 from dandy.bot.bot import Bot
-from dandy.bot.llm_bot import LlmBot