+++ dandy/intel/intel.py
@@ -3,3 +3,5 @@ from pydantic import BaseModel
 
 class Intel(BaseModel):
     pass
+
+